4.TM-Training/requirements.txt
bertopic==0.16.2
cuml==0.6.1.post1
gensim==4.3.2
hdbscan==0.8.37
matplotlib==3.9.0
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.2.2
scipy==1.10.1
sentence_transformers==3.0.1
sparse_dot_topn==0.3.6
